pyproject.toml

[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Adam Homicz <ajhomicz@gmail.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "google-auth>=2.23.0",
    "google-api-python-client>=2.100.0",
]

